making progress on refactor
