+++ src/main.rs
@@ -1,6 +1,7 @@
-mod lib;
+mod monster;
+use crate::monster::Monster;
+use crate::monster::{head::Head, body::Body, arms::Arms, legs::Legs};
 use nannou::prelude::*;
-use the_monster::{Body, Head, Arms, Legs, Monster};
 
 fn main() {
     nannou::sketch(view).run();
